+++ src/test/kotlin/net/pwall/json/JSONSerializerTest.kt
@@ -50,7 +50,6 @@ import java.time.ZoneOffset
 import java.time.ZonedDateTime
 import java.util.BitSet
 import java.util.Calendar
-import java.util.Enumeration
 import java.util.TimeZone
 import java.util.UUID
 
@@ -217,16 +216,6 @@ class JSONSerializerTest {
             addValue("Kotlin")
         }
         expect(expected) { JSONSerializer.serialize(ListEnum(list)) }
-    }
-
-    class ListEnum<T>(private val list: List<T>) : Enumeration<T> {
-
-        private var index = 0
-
-        override fun hasMoreElements(): Boolean = index < list.size
-
-        override fun nextElement(): T = if (hasMoreElements()) list[index++] else throw NoSuchElementException()
-
     }
 
     @Test fun `List of String should return JSONArray`() {